Fixed a win32 watcom project file.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@21389 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/vsyncarch.c
@@ -117,7 +117,7 @@ void vsyncarch_init(void)
             return;
         }
 #ifdef HAS_LONGLONG_INTEGER
-        while (li.QuadPart & (unsigned long long)0xffffffffe0000000) {
+        while (li.QuadPart & (ULONGLONG)0xffffffffe0000000) {
             li.QuadPart >>= 1;
             perf_rotate++;
         }
